// src/schemas/interfaces/LanguageListSchema.ts
export interface LanguageListSchema {
    id: number;
    code: string;
    name: string;
    greeting: string;
    /** Format: uri */
    flag?: string;
    /** Format: uri */
    flagCircular?: string;
    flagEmoji?: string;
    isSupported: boolean;
    levelThresholds: {
        beginner1: number;
        beginner2: number;
        intermediate1: number;
        intermediate2: number;
        advanced1: number;
        advanced2: number;
    };
    learnersCount: number;
}